refactor: move webconnectivitylte algos to webconnectivityalgo
This diff refactors webconnectivitylte by moving some algorithms inside
the new webconnectivityalgo package.

In subsequent commits, we'll seize the opportunity of adding tests for
these algorithms, refactor the code, and add specific tests.

Part of https://github.com/ooni/probe/issues/2669.

While there, recognize that the webconnectivityqa package does not
belong to internal/experiment but to internal.
