+++ internal/experiment/webconnectivity/qa_test.go
@@ -3,7 +3,7 @@ package webconnectivity
 import (
 	"testing"
 
-	"github.com/ooni/probe-cli/v3/internal/experiment/webconnectivityqa"
+	"github.com/ooni/probe-cli/v3/internal/webconnectivityqa"
 )
 
 func TestQA(t *testing.T) {
