+++ internal/webconnectivityalgo/doc.go
@@ -1,0 +1,2 @@
+// Package webconnectivityalgo contains Web Connectivity algorithms.
+package webconnectivityalgo
